--- fourtoon-cookie/src/components/common/Button/Button.styled.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    button: {
-        padding: 10,
-        backgroundColor: '#007bff',
-        borderRadius: 5,
-        alignItems: 'center'
-    },
-    buttonText: {
-        color: '#fff',
-        fontSize: 16,
-    }
-});
